Rename repository and package from ngv-ibkr-reports to ngv-reports-ibkr
This commit renames the entire project from ngv-ibkr-reports to ngv-reports-ibkr:
- Renamed Python package directory: ngv_ibkr_reports → ngv_reports_ibkr
- Updated all Python import statements across the codebase
- Updated pyproject.toml with new package name and GitHub URLs
- Updated documentation (README.md, docs/api.rst)
- Updated build scripts (Makefile)
- Updated CHANGELOG.md with new repository URLs
- Tests passing with new package structure
