+++ CHANGELOG.md
@@ -1,15 +1,15 @@
 # Changelog
 
-## [0.2.0](https://github.com/westonplatter/ngv-ibkr-reports/compare/0.1.1...v0.2.0) (2025-11-14)
+## [0.2.0](https://github.com/westonplatter/ngv-reports-ibkr/compare/0.1.1...v0.2.0) (2025-11-14)
 
 
 ### Features
 
-* add 3.11-3.14 versions ([#16](https://github.com/westonplatter/ngv-ibkr-reports/issues/16)) ([5c0735e](https://github.com/westonplatter/ngv-ibkr-reports/commit/5c0735ecd33cb3bcb8b539aecfc226e4c8c6853a))
-* switch to Python-specific release-please workflow ([#22](https://github.com/westonplatter/ngv-ibkr-reports/issues/22)) ([5e6108c](https://github.com/westonplatter/ngv-ibkr-reports/commit/5e6108c9b46a232381bc7fa8b9d94ca105e75c0d))
-* update to &gt;= 3.9 ([#19](https://github.com/westonplatter/ngv-ibkr-reports/issues/19)) ([cb4ea15](https://github.com/westonplatter/ngv-ibkr-reports/commit/cb4ea1509119349c36f7e45d8f423ba26dc54a6b))
+* add 3.11-3.14 versions ([#16](https://github.com/westonplatter/ngv-reports-ibkr/issues/16)) ([5c0735e](https://github.com/westonplatter/ngv-reports-ibkr/commit/5c0735ecd33cb3bcb8b539aecfc226e4c8c6853a))
+* switch to Python-specific release-please workflow ([#22](https://github.com/westonplatter/ngv-reports-ibkr/issues/22)) ([5e6108c](https://github.com/westonplatter/ngv-reports-ibkr/commit/5e6108c9b46a232381bc7fa8b9d94ca105e75c0d))
+* update to &gt;= 3.9 ([#19](https://github.com/westonplatter/ngv-reports-ibkr/issues/19)) ([cb4ea15](https://github.com/westonplatter/ngv-reports-ibkr/commit/cb4ea1509119349c36f7e45d8f423ba26dc54a6b))
 
 
 ### Bug Fixes
 
-* update docs workflow to use ubuntu-latest runner ([#21](https://github.com/westonplatter/ngv-ibkr-reports/issues/21)) ([cd38570](https://github.com/westonplatter/ngv-ibkr-reports/commit/cd38570f2fd5f6f61c1b14ae4af1e8322f5383c0))
+* update docs workflow to use ubuntu-latest runner ([#21](https://github.com/westonplatter/ngv-reports-ibkr/issues/21)) ([cd38570](https://github.com/westonplatter/ngv-reports-ibkr/commit/cd38570f2fd5f6f61c1b14ae4af1e8322f5383c0))
