+++ ngv_reports_ibkr/adapters.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel
 from pytz import timezone
 from typing import List
 
-from ngv_ibkr_reports.custom_flex_report import (
+from ngv_reports_ibkr.custom_flex_report import (
     CustomFlexReport,
     parse_date_series,
 )
