+++ tests/test_download_trades.py
@@ -1,4 +1,4 @@
-from ngv_ibkr_reports.config_helpers import get_ib_json
+from ngv_reports_ibkr.config_helpers import get_ib_json
 
 
 def test_get_ib_json():
